requirements.txt
matplotlib
networkx
python-constraint
deap
